+++ ui/web/pages/Types/ConnectionTypes.tsx
