+++ ui/web/pages/Types/SuggestionTypes.tsx
